Fix asmc execution in Docker container
When building in Docker, /dev/shm is mounted with noexec, i.e., one
cannot execute files there even if they're executable.

As asmc is downloaded into the build dir which, if possible, is put into
the tmpfs on /dev/shm unless the script is run in CI, we also need to
enable CI mode in Docker.